Game containing code for a demo implementation
I have commented out the code that is currently not working, which I have been trying to get working for a demo implementation of the game. Because this code is commented out, the game still functions as it will have done in the previous version (the version with the fully functioning high-score table).